Add new features and refactor code
Users can now specify in the javascript which branches to target when
creating a PR.

When creating release branches the manager will display the default
branch of the repository and will to base all new release branches on the
default branch.

Created an in memory map of repository meta data for use in the code.

Other minor changes.